request: change parameters type
By make Request::Parameters a typedef for deque<Value> instead of vector<Value>
there is no type ambiguity between a function taking const Request::Parameters&
and one taking const std::vector<Value>&. For the second one the parameters
should be unpacked, i.e. it is the first parameter that is an array.