json: add json format handler
Value has been changed to store numbers as both double, int32 and int64 to be
able to return small int64 numbers as int32.

Also, the binary type has been removed from Value as it doesn't exist in
JSON. When creating a Value from a std::string it is possible to indicate if
the value is binary or not.